+++ lib/scrum/backlog_mover.rb
@@ -5,7 +5,7 @@ module Scrum
     def move(planning_board_id, sprint_board_id)
       setup(planning_board_id, sprint_board_id)
       inspect_backlog
-      @sprint_board.place_seabed(@seabed_card)
+      @sprint_board.place_seabed(@seabed_card) if @seabed_card
     end
 
     private
@@ -18,21 +18,16 @@ module Scrum
       fail "backlog list not found on planning board" unless @planning_board.backlog_list
 
       @waterline_card = @planning_board.waterline_card
-      fail "backlog list on planning board is missing waterline or seabed card"  unless @waterline_card
-
       @seabed_card = @planning_board.seabed_card
-      fail "backlog list on planning board is missing waterline or seabed card"  unless @seabed_card
     end
 
     def inspect_backlog
       @planning_board.backlog_cards.each do |card|
-        if card == @seabed_card
+        if @seabed_card && card == @seabed_card
           break
-
-        elsif card == @waterline_card
+        elsif @waterline_card && card == @waterline_card
           @sprint_board.place_waterline(@waterline_card)
           puts "under the waterline"
-
         else
           move_sprint_card(card) unless @planning_board.sticky?(card)
         end